\name{ds2.mat}
\alias{ds2.mat}
\docType{data}
\title{
    Microarray gene expression data set
}
\description{
    microarray expression data set derived from 5 tissue types (lung, liver, heart, kidney, and brain) from two GEO Series GSE1133 and GSE2361. Each tissue type is represented by 3 replicates.

}
\usage{data(ds2.mat)}
\format{
  A \code{\link{matrix}} with 22283 probe sets and 15 samples.

}
\value{
    microarray data matrix 

}
\details{
    The data consist of the following samples: GSM44702, GSM18953, GSM18954, GSM44704, GSM18949, GSM18950, GSM44690, GSM18921, GSM18922, GSM44675, GSM18955, GSM18956, GSM44671, GSM18951, GSM18952

}
\examples{
data(ds2.mat)
}
\keyword{dataset}
