\name{MGFM-package}
\alias{MGFM-package}
\alias{MGFM}
\docType{package}
\title{
Marker Gene Finder in Microarray gene expression data
}
\description{
The package is designed to detect marker genes from microarray gene expression data sets
}
\details{
\tabular{ll}{
Package: \tab MGFM\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr 
Date: \tab 2014-08-13\cr
License: \tab GPL-3 \cr
}

}
\author{
Khadija El Amrani
Maintainer: Khadija El Amrani <khadija.el-amrani@charite.de>
}

\keyword{ package }


